refactor : client, server instance 하나의 파일로 정리
